KAL-10 Fetch search result with channel names

+++ backend/api/src/main/java/ai/kalico/api/service/youtubej/extractor/ExtractorImpl.java
@@ -144,10 +144,10 @@ public class ExtractorImpl implements Extractor {
     }
 
     @Override
-    public int extractIntegerFromText(String text) {
+    public long extractIntegerFromText(String text) {
         Matcher matcher = TEXT_NUMBER_REGEX.matcher(text);
         if (matcher.find()) {
-            return Integer.parseInt(matcher.group(0).replaceAll("[, ']", ""));
+            return Long.parseLong(matcher.group(0).replaceAll("[, ']", ""));
         }
         return 0;
     }
